MBS-12722: Add IdRef to the other databases whitelist
This is basically a French authority control page,
kind of ISNI-like.

Allowing the entity types yvanzo and I found examples of so far.
Since the URL patterns are the same for all of them,
I didn't bother adding equivalent examples for all types.

There's also links for areas, but we don't currently support
otherdatabases links for areas at all, and those are only edited
by a couple people anyway, so probably not worth adding the
relationship type just for this.