Document and improve Entity::AutoEditorElection test
This documents the test properly and fixes the style.
The test itself was a stub of dubious usefulness,
so I added tests for several methods in the package as well.
I changed this to use has_attribute_ok rather than can_ok
since that is specifically meant for Moose attributes,
which is what we are checking here.