Don't apply both Role::SelectAll and Role::OptionsTree
Role::OptionsTree already applies Role::SelectAll,
so trying to apply it again causes Moose issues
(these are currently hidden because we incorrectly use multiple with).

Since we want to be able to pass an order_by parameter to SelectAll,
I made OptionsTree parameterizable and made it pass the parameter
through.