Avoid using multiple with in EditSearch:: files
https://metacpan.org/pod/Perl::Critic::Policy::Moose::ProhibitMultipleWiths

Since this involves a lot of files, doing it in parts to ensure
it's easier to find what broke if something does -
this just for EditSearch::

Predicate and Predicate::Role::User both implemented a valid method,
causing a conflict. Since roles are meant to be used with Predicate,
changed the method to an around that uses the base method too,
which seems closer to what was intended anyway.