Refactor: Split get_by_ids in Data::ArtistCredit
Split the method `get_by_ids` of `Data::ArtistCredit`
(which overwrites the same method of `Data::Entity` it extends),
so that the second part can be used for getting ACs by GID too.

The new subroutine for this second part has the distinctive name
`_get_hash_by_keys` because it doesn’t return an array like
`_get_by_keys` of `Data::Entity` does but it is the same otherwise.

Update the POD accordingly.