Overwrite _get_by_keys of Entity in ArtistCredit
In `Data::ArtistCredit`, the method `_get_by_keys` was inherited from
`Data::Entity` but wasn’t used so far and wouldn’t work as-is.

This patch overrides this method so that it works as it is required
by the role `GetByGID` added to `Data::ArtistCredit` in commit
8a841d62b5454cdb5bcc2f621ab6b7791567ddc6 without making use of it.

Update the POD accordingly.