MBS-12367 (JSON): Add artist-credit-id to output
Since `'artist-credit'` is a JSON array of artist credit names, adding
artist credit’s MBID under it would require to turn it into an object
which would be a breaking change for everything that uses the WS/2 API.

Instead this patch serialize artist credit’s MBID and add it under the
new key `'artist-credit-id'` at the same level as `artist-credit`.

Tests are updated in a separate commit.