Add separate serialize_artist_credit method
This avoids issues with the current AC serializer and serialize_entity
and would be used later anyway for serializing artist-credit-id
and adding it to the API.

As such, I've cherry-picked it from yvanzo's implementation of MBS-12367
and just commented out the artist-credit-id section for now.

Since `'artist-credit'` is a JSON array of artist credit names, adding
artist credit’s MBID under it would require to turn it into an object
which would be a breaking change for everything that uses the WS/2 API.

Instead MBS-12367 will serialize artist credit’s MBID and add it under the
new key `'artist-credit-id'` at the same level as `artist-credit`.

Tests for id will be added during MBS-12367 implementation.