Document Data::ISWC test
This documents the test properly (most of it was already fine)
and improves comments a bit.
I got rid of the seemingly unnecessary separation of tests
into smaller blocks, which we don't generally do elsewhere.
Also, the find_by_iswc test was testing the Data::Work method,
not the Data::ISWC one, which seems unintended, so I fixed it.