MBS-12778: Don't block area removal based on open edits
For some reason we were checking for open edits on can_delete for areas.
This is what we do for entities that are autoremoved when empty,
but areas are not one of those. In fact, this being here was blocking
area removals entirely AFAICT, since any removal edit would be rejected
because that same edit was open at the time of the can_delete check.