MBS-12780: Fix allowNew check for AddRelationship
This was checking for the entity0 and entity1 passed outside
the relationship itself - which we were also adding
to the relationship data for display.
But we were only passing this if the ID was not defined
in the first place, so this check was nonsensical AFAICT.
In fact, I cannot find any reason to pass entity0 and entity1
outside the relationship, since we're already passing them inside it.

I'm also removing the source_type and target_type that
I had added in 9b0eae3e837e848c5eecb71b574185a37a2e9387
since AFAICT we're again not using those, but the ones
inside the relationship object.