MBS-12775: Convert the release Disc Ids page to React
Since the release editor expects medium.cdtocs to be just
an array of TOC strings, and this is the only place we seem to use
the full MediumCDTOC, I am passing the full MediumCDTOCs
as an array via props.
Since we already load all the needed data for isPerfectMatch
as part of that array, I am using the medium data in there
rather than on the release mediums list to build the rows.