MBS-12741: Convert Create/Edit attribute admin pages to React
TT was happy to have a scrap heap of forms held together with wire
and glue here, but I thought it would make a lot more sense to split
the different forms into their own components and pick the right one
according to the model.

I made it so that the entity type dropdown actually shows
translatable entity names, rather than the type string.