MBS-12785: Wrap things sensibly on release sidebar
The release sidebar has two places where an entity name
in EntityLink can be huge and cause the sidebar to grow
to fairly unreasonable sizes:
the label name (under Labels) and the release group name
(for the CritiqueBrainz RG reviews).

This makes sure those can always wrap as needed; rather than adding
a third class name to do the same, I just renamed the existing ones
to a very explicit "wrap-anywhere".

Other entity links on the sidebar are controlled by us
(language and script names, release countries)
so probably do not risk being a problem with this,
and tags already wrap.