MBS-12753: Convert edit type documentation page to React
Much of the info in the table for this page is no longer relevant:
voting period is now always 7, needed votes always 3,
and every edit is accepted on expiration.
As such, I dropped those rows, and I'm adding a row for whether
an edit type can be automatically applied even by normal editors
(such as Add artist), by auto-editors only (such as Add cover art)
or not at all. I added a new condition for this, which is false
unless set to true by Role::AlwaysAutoEdit.