MBS-12801: Warn users about "Disc n" titles on release merge
This is a basic warning, which is probably enough. We could also disable
submission unless they confirm, by erroring at the Perl level, but this
is less intrusive and hopefully does the trick still, since it appears
just below the name row as soon as the user enters the triggering name.