MBS-12809: Pass minimal existing data for Historic/RemoveDiscID
We were trying to build a new CDTOC from the discid alone,
which actually goes against our definition of a CDTOC entity.
This was now breaking in TO_JSON since it now expects a track_offset.

Rather than making that optional everywhere, I'm just passing
a "minimal-cdtoc" JSON object with the data we have
and allowing CDTocLink to use it - it does not need anything else,
in any case.