MBS-12815: Support immutable URL edit search predicate
Relationships involving URLs have an edit search link when
the relationship has pending edits, like for any other entities.
Problem is, we do not intentionally support URL as a predicate at all.

In fact, this search was working, because Entity also created an
(undocumented) predicate for URL, but this had no label
and a search field.
URLs are not searchable, so the search being there was misleading,
and having no label leads to it being listed as ID.

I decided to implement a basic, uneditable option to properly support
this use case, which shows the URL as text.
I made it support "is not" as well because is trivial,
although I'm not sure if it's useful tbh.

This is only shown if URL is being seeded as an option to begin with;
it's not selectable for users.

This is all a big hack that should be replaced by MBS-12122
but it feels better than the current situation until then.