MBS-12816: Highlight doc header when linking to fragment
It's often unclear where a direct link to a fragment is pointing
when opening a doc page, especially when the page
is short enough the browser does not load the linked section
at the top of the window.

This at least highlights the header for that section;
in order to highlight more than that we'd probably need
to either change the HTML to have the fragment ids on a <section>
surrounding the heading + paragraphs, or add a class directly
to all paragraphs after getting the HTML from the wiki.