MBS-12787: Editing a relationship to add instruments causes duplication
If a "members" relationship is added without any instruments, submitted,
and then edited to add instruments, then the relationship will be
duplicated for each instrument (or vocal) instead of adding at least one
to the existing relationship.  ("members" is just an example, as this
could apply to any relationship type supporting instruments or vocals.)

The solution is, if the existing relationship has no instruments or
vocals, to take the first new instrument or vocal that would otherwise
be split, and merge it into the existing relationship's attributes.

The "first" instrument or vocal chosen is the one with the lowest link
attribute type ID, hence why the added test expects vocals to be chosen
for merging over drums.