MBS-12806: Do not show "Play on LB" button for empty releases
If a release has no tracks at all, there's no point trying to play it,
so we might as well not show the LB button on the sidebar
and annoy the user who tries.

AFAICT there's no trivial existing method to do this,
so added a new has_no_tracks method.