MBS-12827: Run our sanitize function on new usernames
This shows an error on the registration form if the name doesn't match
the sanitized version of the name.  I imagine this will mostly be
accidentally triggered by people adding extraneous whitespace
characters.

If an unsanitized username somehow makes it past our form validation,
the editor data layer performs the same check and possibly dies.