MBS-12831: Don't hide disambiguations in relationship previews
I'm not sure why this was actively disabled here (the commit history
shows nothing about it) but it certainly seems useful to see
the disambiguations in the preview. There's at least two cases
where this can be very useful:
1) When the two entities have the same or a very similar name,
so the disambiguation is the main way to understand the selected order.
2) When the editor misclicked and picked another entity
with the same name as the intended one, having the disambiguation here
makes it easier to notice and fix immediately.