Don't close relationship dialogs when clicking outside of them
This was reported in the comments of MBS-11847 as being a potential problem
with respect to losing edits.  On IRC it was also decided that being able to
copy text or links from outside of the dialog without closing it may be useful.

Note: this restores the current behavior in production.