Autocomplete2: Mimic jQuery UI Autocomplete style
There's been some discussion about the orange menu highlighting being
confusing in MBS-11847.  I tried changing it to green as suggested, but
others thought the green was weird.

Try to mimic the old jQuery UI autocomplete menu highlight style, since
nobody seemed to have an issue with that.  (It may not be a perfect
reproduction.)