MBS-10609: Actually check for edit notes by beginners
AFAICT this was just ignoring the fact this was meant
to be an edit note search and just searching for edits by beginners,
since the entire sql was replaced when type is limited regardless
of the passed clause.
mwiencek suggested the current way for us to be able to still use
the passed template clauses, which seems to work nicely.