MBS-12834: Suppress relationship type menu on dialog open
When an Autocomplete2 input is focused, it normally shows its menu.  If
the link type is automatically focused when the dialog for an existing
relationship opens, then it will also show its menu right away, which
interferes with editing.

We now only show a menu on focus if there's no existing selection.

It was also pointed out in the ticket that it's inconsistent which input is
focused if the source entity is creditable.  We now focus elements in this
order where possible: target entity type, then link type.  This is controlled
via a 'focus-first' CSS class.  If multiple elements have this class, then the
first one having it is focused.

Because having the target type initially focused isn't very useful when
you click one of the "Add another {entity}" icons (which is intended for
adding another relationship of the same type), I've made the target type
static in this case.  The option to change the target type is now only
available when clicking on the general "Add relationship" buttons.