Revert "Autocomplete2: Don't update menu on mouse move"
This reverts commit ab7ad235b39e4a24717102168b2018d2d8509be3.

It was mentioned in IRC that having two menu items highlighted (one for
keyboard focus, another for mouseover) is confusing.  The performance argument
doesn't justify having confusing behavior, especially since the extra latency
is barely noticeable.  The comment in the reverted commit about React having to
re-render the entire component if the mouse moves even 1px also isn't true --
it re-renders on mouseenter, not mousemove.