Also show disambiguation if there's an entity credit
It is useful to always show disambiguations on the relationship preview,
even when one of the entities has a credit (for some reason, EntityLink
defaults to hiding them in that case unless explicitly requested).
This is for the same reason why we stopped *blocking* showing
disambiguations here: you can have two entities of the same or similar
names on both sides of the relationship, and thus have given the credit
to the wrong one by mistake, which is hard to notice
if the disambiguation is not shown.