MBS-12860: Fix "Two relationships with the same key" error
This reverts commit 6230a40b799f0879408fe03d79d8fce83293c0d3.

That commit caused new relationship IDs to become reused by memoizing
the initial relationship data seeded to dialogs.  It can be safely
reverted in full because it was an unnecessary optimization in the first
place.

You can trigger the reported error by adding two area relationships to
any entity in the release relationship editor.  You can trigger it for
target types other than area too, but the catch is that the memoized ID
only changes when the target type changes, *after* the dialog is closed.
So to trigger it for artist relationships from a fresh page, you'd have
to add three artist relationships to the same entity.

While I can't be 100% certain this is the only cause of the key error,
it seems pretty likely that it's the most common one.