MBS-12874: Don't split new relationships with one instrument or vocal
When you edit a relationship, we look at all the instruments and vocals
added and try to split each one into a separate relationship.  That's
handled by `splitRelationshipByAttributes`.

If the relationship being edited is new, i.e. not submitted yet, then
prior to this commit we would incorrectly "split" this relationship into
a new one (a clone) even if it only had a single instrument or vocal
attribute.  The correct course of action is to simply return the same
relationship back.

One side effect of not doing that was having triggered an error where
the same relationship state is tried to be removed twice: once to
replace the existing relationship state as edited by the dialog, and
another to remove it again when the bogus "split" relationship (which is
identical to the original other than by ID) is merged into the original
by `mergeRelationship`.  The second removal fails because the
relationship no longer exists at that point.