MBS-12879: Block Bandcamp /videoframe links
We got a report that we were mangling Bandcamp /videoframe links,
since the cleanup was removing the video ID from them.

On researching it, yvanzo found that these seem to be intended
as Bandcamp's inner working links, as they are, as far as
we have found, meant to be shown in an iframe in /track pages
(notice that the links are not even listed under the appropriate
artist.bandcamp.com subpage).

As such, this blocks them (without breaking them)
and suggests users to use the appropriate /track link instead.