Widen the combination options for Soundcloud URLs
This was originally limited to free+free and paid+paid
because that's what we found, but we didn't check exhaustively
and the main reason we limited it like this IIRC was because the list
seemed huge in the code already. The original request was just
to block it to only combinations of "get the music" rels.

I'm extending this to all possible combinations as a proof of concept
for a new anyCombinationOf function that can also be used
for other similar cases, such as Bandcamp.

The basic code for generating the combinations was taken from
https://stackoverflow.com/questions/71878018/javascript-all-possible-combinations-from-single-array-every-order