MBS-12885: Instrument disappears after adding another
This is a similar issue to MBS-12874 in that the code is still trying to
merge a relationship into one that no longer exists.  In fact, the fix
in this commit fixes both issues.  (The initial fix to MBS-12874 still
makes sense too, though.)

Once a relationship dialog is accepted, we should make sure to remove
the old relationship state before processing any other updates,
especially ones that attempt to merge the new relationship state;
otherwise it's very possible that it will find the relationship state we
are about to remove and select it as a merge target.