MBS-12888: Check ended when copying end date
This is just a visual issue because in the backend adding an end date
automatically sets ended, but it still makes sense IMO that this would
be visible on the form as well.

Flow wants the PartialDateStringsT cast to be sure
which of the two types it's supposed to look at.