MBS-12899: Add TMDB to the otherdbs whitelist for artists
This is for film (no music videos) but it has a lot of artists
who are also in MB and biographical data for them.
Several editors asked to be able to link to this,
so I guess we can.

For now allowing linking only to artists since it seems like
the only clear match. Added a test to check linking films to works
is blocked.