Autocomplete2: Wait for recent items before showing menu
In Autocomplete2, the recent entities list is stored as a list of MBIDs
and then fetched asynchronously. This was done to resolve issues like
MBS-10498, and generally to prevent the cached entity JSON from becoming
outdated.

This creates a race condition where if the field is focused before the
recent entities are loaded, they won't be displayed.

To fix this, in this commit I store the request promise in a React ref,
and if we attempt to show available menu items, wait for that promise to
resolve first.