MBS-12901: Show user website an bio on the admin user views
The bio is very useful to figure out at a glance
whether someone is a spammer (no bio doesn't mean they are not,
but spammy bio automatically means they are).
As such, it would be great to see on the IP lookup results, at least.
Since all admin user lists use the same component, and it's not like
the bio *hurts* in any way elsewhere, this will also show them
on the email and privilege search tools.
In some cases, the bio is empty but there's a clearly spammy
website, so also listing that.
This brings the useful data shown into parity with the
Beginner editors report, which makes spammer finding very easy
in my experience.