MBS-12873: Avoid unloaded tracks to change credits
In the release relationship editor, not all tracks of big releases are
initially loaded, therefore the option to change credits for all other
relationships isn’t working as expected in this case.

This patch disables this option when not all tracks have been loaded,
and adds an help message to work around this current limitation.

It makes the release relationship editor to re-render for big releases,
only once the last batch of tracks has been loaded.

Existing test has been updated to make it pass.
There is no automated test for this workaround.

The long-term solution should be to load all tracks when ticking it.