Fix selecting mediums with no recordings
In the release relationship editor, a medium can have no recordings
if it is just not loaded yet, for big releases. In this case, the
medium was misleadingly selected by default for batch-adding
relationship to recordings. It mixes up with MBS-12875.

This patch fixes that by using the same approach as for linked works.