MBS-12875: Disable selecting incomplete mediums
In the release relationship editor, for big releases, a medium that is
not fully loaded or not loaded at all should not be selectable for batch
operations on its recordings as these are not yet loaded. For the same
reason, all mediums cannot be selected together in this case.

This patch fixes that by using the same approach as for MBS-12873:
disable the checkbox and add an help message when appropriate.