MBS-12904: Fix another error submitting a recording artist relationship
This issue is very similar to MBS-12694, but involves selecting the
recording artist from a search first in order to push it onto the recent
entities list.

The issue was that, when displaying search results, jQuery UI would
modify the entity JSON to add undefined `label` and `value` properties
as part of its "normalization" process. These undefined properties would
cause `compactEntityJson` to throw an error.

I removed the normalization code from jQuery UI because its behavior is
gross and we don't appear to need it anywhere. We override `_renderItem`
everywhere, and don't require `label` or `value` to display anything
that I can see.