Catch prepareHtmlFormSubmission errors
Prior to this commit, if an error were to occur inside
prepareHtmlFormSubmission, then the submission would still proceed
without showing any information to the user or even logging the error to
Sentry. Now we catch such errors and additionally cancel the submit
event in order to inform the user that something went wrong.