MBS-12829: Use link type phrases in autocomplete input value (#2840)
I decided not to change the input value based on the direction of the
relationship, because the forward/reverse link phrases aren't meant to
be read in sentence order like the long link phrases are, so I don't
think it makes the selection easier to understand. It's also harder to
implement.