MBS-12665: Widen DNB regex (#2686)
We took this from Wikidata, but given the valid example added
as a test here, it was too strict.

phw/outsidecontext got in touch with DNB and got a kinda not
very useful answer, but his proposal based on it was to just
allow digits plus X and hyphen. It seems that the length
should be either 9 or 10 characters. Further assumptions
seem a bit dangerous so I'm not making any.

The understanding is that there's two main kinds of links.
One is the /gnd type, which is for their integrated authority file
(a catalogue of people, organizations, works and a few other
entities), and the other is the type without a subpath,
which is supposed to be bibliographic data and the only one
where releases are stored.
A third sort of link, /dnbn, seems to be used for publishers
and is currently made available for labels. I'm not sure
if self-published artists also have publisher pages, but
we'd probably prefer linking to their actual artist page anyway.