MBS-12798: Cancel user's open edits when marked as spammer
We used to delete spammers, but now we leave them behind
(marked as spammers) so that SpamBrainz can eventually learn from them.

When we delete an editor, their open edits get cancelled and
their votes on open edits changed to Abstain.
It seems sensible to do the same for spammers, since
a) we used to remove them
b) their edits and votes are, by definition, dodgy.