MBS-12921: Can't search for works in the "Add related work" dialog
The Autocomplete2 component was attempting to show recent items instead
of performing a search, because of an incorrect check where it assumed
recent items would show even if the input was non-empty (they won't).

The reason the issue only manifests in the "Add related work" dialog is
because we pre-fill the input value with the recording name there.

I've also changed the set-menu-visibility action to only open the menu
if there are actually items; prior to this commit you could see an empty
"menu" open (i.e. a gray line) beneath the input, and hitting arrow-down
to move the item position would throw an exception because there are no
items to highlight.