MBS-12926: Remove indents from Autocomplete search results
Level indents make sense when displaying the whole set of results
as a static list (since then you can see that, say,
"instrument" is a child of "performer").
When filtering the list, if "instrument" is still indented, it will look
like it's a child of whatever happened to be on top of it,
which is at best useless and at worst makes the user think it's
really connected to the previous result in the list.

We already remove the indent level for recent items for the same reason.