MBS-12930: Validate relationship fields on Perl level
We were showing an error for this on the JS level, but
(because the submit button is a buggy knockout mess and wasn't disabled)
it was still possible to try to submit a seeded relationship
without selecting a type, which would cause an ISE.
Now it properly detects that in fact something is missing and blocks
it from submitting.

I don't think the error here is currently being shown on the browser,
but since it's the same string as for the JS one, it should be fine.