MBS-12925: Allow submitting work dialogs with Enter
Other dialogs in the release relationship editor can be submitted
with Enter without issues, but that was not the case for the work
dialogs (batch-create and edit).

This follows the example of RelationshipDialogContent by submitting
on pressing Enter unless it is pressed while on a button
(to avoid an Enter press on "Cancel" submitting the form).

I turned the batch-create fragment into a div to support onKeyDown;
I added the "form" class because it seemed sensible for parity.