MBS-12919: Show warning icon when rel has pending edits
We used to show (previous) pending edits by highlighting
the entity in orange, but that's now how changes made
during the active editor session are shown so that doesn't work anymore.
I decided to show a small warning sign, with the same pending edits link
on hover as we use once you in the relationship editing popup.