Throw error if nameVariation is not set as needed
When the content of a link to an entity is a React component, it cannot
be compared as a string to entity name, then nameVariation must be set.

Spots any misuse with an error message sent to the developer,
either by crashing the page in development setup or through Sentry.

Follows commit 8fcf10454f98090f8bbd51aa992340b321e3bf17