MBS-12622: Show if URL rel has pending edits in links editor
We now show a special icon on the (non-URL) relationship editor
if a relationship that can be changed has (previous) pending edits.

This adds the same icon to the URL editor, where nothing
was being shown currently. This should help avoid unnecessary
URL removals when someone is already setting it as ended, for example.

The externalLinks code does slightly odd things to entity/relationship
data to put them into state, so flow typing here is pretty messy -
unsure if it can be improved without a proper refactoring
of the externalLinks component though.