MBS-12936: Don't update Autocomplete2 scroll position on item mouseover
In d4b15e9de756331755850ee6fb8479c0f42be14b I had intended to fix an issue
where highlighting an item lower down in the list (thereby updating the scroll
position), closing the menu (e.g. by hitting escape), and then re-opening the
menu would not bring the scroll position back to the top of the list.  While I
did fix that, there were some unintended effects when using a mouse: hovering
over any item would cause the menu to scroll towards that item.

The actual intent is that we should only update the scroll position when using
the arrow keys or when opening the menu. Obviously if a user can mouseover an
item, then it's already in view, and we don't need to mess with their manual
scrolling.  This commit fixes that and adds a Selenium test to make sure we
don't regress in this behavior again.