MBS-12935: Readd relationship type help in editor
Since the autocomplete menu doesn’t show if there is a focused item,
the relationship type’s description is no longer always visible.

This restores the description that was displayed under the selector in
the previous (TT/jQuery) implementation of the relationship editor.

It also readds the “(help)” toggle button but places it in the new
“Relationship” headline instead of at the right of the selector,
so as to follow the same style as the “Attributes” headline.