Merge help toggle links in relationship editor
The problem with having “(help)” toggle links in the headlines of
section “Attributes” and “Relationship” is that it requires editors
using keyboard navigation to press the Tab key two times to skip these.
Another issue is that if you want to see the help for both relationship
type and attributes (which actually depend on the relationship type),
editors have to click two times instead of one.

This patch merge the two toggle links into one and place it in the top
left corner of the relationship editor dialog. It keeps using “(help)”
text for now, which is already in use in other parts of the editing UI.

In the longer term, this link could be replaced with a fancy CSS switch.

Since the focus is set on the “Related type” field at first, this link
does not have to be skipped in keyboard navigation, but it stays
accessible - just like the source entity link - by back-tabbing or
cycling through the whole dialog.