MBS-12931: Hide (+) button if no relationship type
When creating a new work for an artist, the relationships group editor is
prefilled with a relationship to the artist, but the relationship type
is not selected. This is up to the user to select it. There is no point
to allow adding a relationship of the same type if it isn’t defined yet,
as there already is a button to add a new relationship of any type.

This issue has been uncovered by invariant error message in development
mode.

This patch just hides the (+) button as long as no relationship type is
selected.