Convert 0 linkTypeIDs to null for the relationship dialog
This fixes the underlying cause of MBS-12931 by converting link type IDs stored
as 0 to null before constructing the relationship state sent to the dialog.

It also adds some Sentry logging in case we see a link type of 0 in
useRelationshipDialogContent again, though I think the previous fix for
MBS-12931 (hiding the '+' button when there is no type) makes that very
unlikely.