Fix invariant violation submitting external links
88a1a97b0709233f1919a217fc33a7fa381a98dc wrapped the ExternalLinksEditor
component in `withLoadedTypeInfo`, which ensured that link_type and
link_attribute_type data are loaded asynchronously before the component
renders.

If the user happened to submit the form before `withLoadedTypeInfo` completed,
then an invariant violation would be triggered due to the component ref being
null.

Since there's no other case where this ref should be null, and since we know
the user didn't intend to submit any link changes (otherwise they would have
waited for the component to finish loading), we can just ignore the situation.
If there was another error that somehow caused the ExternalLinksEditor to
become unmounted, it would already be logged to Sentry (but I didn't find any
such error).