MBS-9063: Don't try to submit dates for reltypes without them
We keep having issues where people will hit an ISE after changing
from a relationship type with dates (such as member) to one without
(such as founder), because the dates are still stored and submitted
with the relationship even though it doesn't make sense.

Since we do not want to lose the dates in a situation where the user
ends up selecting a second type that does support them,
this stashes and blanks the dates if a type without dates is selected,
and restores the dates from the stashed ones if a type with dates
is selected again.