MBS-12946: Don't put Tooltip div inside span
ArtistCreditLink was trying to put the Tooltip <div> inside a <span>.
This is invalid HTML, and React was not happy, so it broke hydration
and looked stupid to boot, with a duplicate pending edits icon.

There seems to be no downside to just making Tooltip use <span>
all the time, other than it needing to be set to display as a block.
As such, this just makes that change.