Upgrade Flow to 0.196.0
https://github.com/facebook/flow/releases/tag/v0.196.0

From the changelog:
"Instead of mixed, type the result of Object.values and Object.entries on a
dictionary to be the dictionary values, and Object.entries keys to behave like
Object.keys."

This only affected one line in linkedEntities, but basically just forced us to
have to change which error was being ignored. The values in this case are still
typed as `mixed` because LinkedEntitiesT is not an indexed type.