MBS-12970: Negate nameVariation for release country diffs
This was being compared with a string even when passing a span,
hitting the invariant error we added for this situation.
We're expected to set nameVariation manually in these cases.
Normally we might set nameVariation to true
since we're not passing the name, but in this case we pass
either the area name or a span containing the area name,
so it probably makes sense for neither of those to be consider
an actual nameVariation,
since I understand the point of that is to help find cases
where a different name is printed.