MBS-11565: Use less hopeful image not available message (#2872)
A lot of the times, the image is not actually going to appear
in a few minutes, but having that in the text makes people think
the longer delay is uncommon and needs reporting.
Sadly, it isn't. This will hopefully limit their optimism and thus limit
the amount of reports we have to answer with "shrug, it's the IA".

Since this message is also shown when images have been processed,
but the IA is down, this also removes the "yet", which is weird
when shown for an image that already worked before.