Upgrade Flow to 0.201.0
https://github.com/facebook/flow/releases/tag/v0.201.0

From the changelog:

"Remove unused-promise-in-async-scope and unused-promise-in-sync-scope. They
can be enabled with unused-promise."

This means we now have to handle all former "unused-promise-in-sync-scope"
errors. I've done this by adding .catch handlers where needed. Where the caught
error involved a network request, I added some display code to show the error
to the user.