Change React.AbstractComponent -> React$AbstractComponent
This supposedly has no downside while allowing us to remove
imports for React in quite a few files.

For some reason, using $AbstractComponent did trigger
nonpolymorphic-type-app flow errors that were not triggered
by .AbstractComponent.
Might be an underlying flow bug?
In any case, adding void or mixed like we already did for a fair amount
of AbstractComponents should not be a problem.