MBS-12984: Link zoomable image in the WikiDocs
Previously, links on images were removed because images didn't have to
be zoomed anyway.  Larger images (such as database schema) were linked
through additional links that had to be added manually in the wiki.

This patch allow linking a thumbnail to the original image simply by
adding the option "class=zoomable" to the wanted "Image" wikilinks.

It's cheaper than MediaWiki's MultiMedia Viewer extension but it works.