MBS-12977: Correctly specify the licenses of the JSON dumps
We were claiming this is all CC0, but that is not correct.
The JSON dumps contain annotation, rating and tag data,
all of which is supplementary / CC-BY-NC-SA data.

Talked with mayhem and mwiencek about this, we agreed
that the right thing to do is to update the dump docs to specify
the right licenses rather than dropping the supplementary data,
since it can be useful and there's no great way to provide it separately
like on the MB dumps.