MBS-12122: Add URL as edit search option
While this is not yet searchable in an autoselect kind of way,
mwiencek suggested the best way to do this for now would be
to turn the field into a searchable text field,
so that the user could enter a URL in it.

This changes getOpenEditsLink to also generate the right style
of link when one of the sides of the relationship is a URL.

This will only match exactly. If the URL has no ending slash
and the user searches with one, or if they search HTTP vs HTTPS,
it will find no results. It still seems like a lot more useful
than nothing, which was what we had before.